third_party/blink/web_tests/platform/generic/http/tests/devtools/elements/edit/insert-node-collapsed-expected.txt

Tests that elements panel updates hasChildren flag upon adding children to collapsed nodes.


Running: testDumpInitial
========= Original ========
  <div id="container"></div>

Running: testAppend
======== Appended =========
+ <div id="container">…</div>
